algebraic.mle: A package for algebraically operating on and generating maximum likelihood estimators from existing maximum likelihood estimators.

The object representing a fitted model is a type of mle object, the maximum likelihood estimator of the model with respect to observed data.

Details

In what follows, we briefly define the API (generic functions, mostly) with default implementations for objects that inherit from mle.

vcov(mle) returns the variance-covariance matrix of the model parameter estimates.

point(mle) returns the point that maximizes the likelihood of the model.

confint(mle) returns confidence intervals of the true parameter value estimated by the mle object.

sampler(mle) maps to a function that may be used to sample from the sampling distribution of the mle object.

mse(mle) computes the asymptotic mean squared error of the mle object.

mle_weighted(mles) computes the weighted maximum likelihood estimate from independent mle objects in the list argument mles.

fisher_info(mle) returns the Fisher information matrix of the model's parameters.

rmap(mle,g) provides numerical_mle the maximum likelihood estimator of g(mle). (Note that numerical_mle inherits from mle.)

mle_linear_transform(A,mle) provides a closed-form solution of the linear Finally, since normal distributions are closed under linear transformations, then letting A be a p-by-q matrix and x be a q dimensional mle object, then prod(A,x) is a p dimensional multivariate normal with a mean prod(A,x) and a variance-covariance prod(tr(A),vcov(x),A. Furthermore, by the invariance property of the MLE, prod(A,x) is the MLE of prod(A,true(x)) where true(x) is the (usually unknown) parameter being estimated by x.

We define a set of generic functions for multiplying mle objects by (non-random) matrices and vectors to permit these sort of operations, with the asymptotic distributions of the operators applied to the mle objects exactly known if the asymptotic distribution of the mle objects are exactly known.