`algebraic.mle`: A package for algebraically operating on and generating maximum likelihood estimators from existing maximum likelihood estimators.

The object representing a fitted model is a type of `mle` object, the maximum likelihood estimator of the model with respect to observed data.

Details

It has a relatively rich API for working with these objects to help you understand your MLE estimator.#'

See also

Author

Maintainer: Alexander Towell lex@metafunctor.com (ORCID)